cleanup the protocol client API
